+++ examples/bop.rb
@@ -8,7 +8,7 @@ ROOM = ENV['ROOM']          # 'xxxxxxxxxxxxxxxxxxxxxxxx'
 
 TT.run(EMAIL, PASSWORD, :room => ROOM) do
   on :user_spoke do |message|
-    if message.text =~ /bop/
+    if message.content =~ /bop/
       client.room.current_song.vote
     end
   end